updated prefabs & removed navmeshobstacles, baked navmesh, fixed UV scrolling with one scrolling script (vs per lava gameobject)
