+++ Assets/BossRoom/Material/Dungeon/lava.mat
@@ -42,7 +42,7 @@ Material:
     - _MainTex:
         m_Texture: {fileID: 2800000, guid: 71f14d5f7422f9346af23185b106a77b, type: 3}
         m_Scale: {x: 1, y: 1}
-        m_Offset: {x: 0.2626989, y: 0.2626989}
+        m_Offset: {x: 0, y: 0}
     - _MetallicGlossMap:
         m_Texture: {fileID: 0}
         m_Scale: {x: 1, y: 1}
