+++ Assets/BossRoom/Prefabs/Dungeon/Dungeon Pieces/animated/env_treasure_chest_ani.prefab
@@ -12,8 +12,7 @@ GameObject:
   - component: {fileID: 445632146649864073}
   - component: {fileID: 7463718357983197496}
   - component: {fileID: 3577683239110096990}
-  - component: {fileID: 5277896896820883330}
-  m_Layer: 0
+  m_Layer: 8
   m_Name: env_treasure_chest_ani
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -96,19 +95,3 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 3.641676, y: 3.8060122, z: 4.927659}
   m_Center: {x: -0.013105154, y: 1.8930116, z: 0.00000023841858}
---- !u!208 &5277896896820883330
-NavMeshObstacle:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4371491365999748502}
-  m_Enabled: 1
-  serializedVersion: 3
-  m_Shape: 1
-  m_Extents: {x: 1.8210001, y: 1.9030001, z: 2.4640002}
-  m_MoveThreshold: 0.1
-  m_Carve: 1
-  m_CarveOnlyStationary: 1
-  m_Center: {x: -0.013, y: 1.8930001, z: 0}
-  m_TimeToStationary: 0.5
