+++ Assets/BossRoom/Prefabs/Dungeon/Dungeon Pieces/env_floor2.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 445632146649864073}
   - component: {fileID: 7463718357983197496}
   - component: {fileID: 6399737381871187658}
+  - component: {fileID: -3983996757204683007}
   m_Layer: 8
   m_Name: env_floor2
   m_TagString: Untagged
@@ -95,3 +96,17 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 4.0000005, y: 1, z: 4.000001}
   m_Center: {x: 1.9999998, y: -0.5, z: -2.0000005}
+--- !u!114 &-3983996757204683007
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4371491365999748502}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8ee75574922a1ab448c29249c9d11e2f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Collider: {fileID: 6399737381871187658}
+  m_EnableStateOnAwake: 0
