+++ Assets/BossRoom/Prefabs/Dungeon/Dungeon Pieces/env_spawn_door.prefab
@@ -236,7 +236,7 @@ GameObject:
   - component: {fileID: 8184890942973508939}
   - component: {fileID: 4897062626467851486}
   - component: {fileID: 8778471712971589033}
-  m_Layer: 0
+  m_Layer: 8
   m_Name: env_spawn_door
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -262,8 +262,6 @@ Transform:
   - {fileID: 1460004792172074384}
   - {fileID: 2777313110340582950}
   - {fileID: 6681280469685477543}
-  - {fileID: 4150610526904978245}
-  - {fileID: 3871733559294641458}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: -90, z: 0}
@@ -452,53 +450,6 @@ MonoBehaviour:
   m_MinIntensity: 0
   m_MaxIntensity: 2
   m_Smoothing: 36
---- !u!1 &2577209187459416071
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 3871733559294641458}
-  - component: {fileID: 3008680131107812324}
-  m_Layer: 0
-  m_Name: Nav Mesh ObstacleL
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &3871733559294641458
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2577209187459416071}
-  m_LocalRotation: {x: 0.000000021855696, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 7980115660496427320}
-  m_RootOrder: 9
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!208 &3008680131107812324
-NavMeshObstacle:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2577209187459416071}
-  m_Enabled: 1
-  serializedVersion: 3
-  m_Shape: 1
-  m_Extents: {x: 1.065, y: 3.4370434, z: 1.29}
-  m_MoveThreshold: 0.1
-  m_Carve: 1
-  m_CarveOnlyStationary: 1
-  m_Center: {x: 0.57, y: 2.187421, z: -0.006958308}
-  m_TimeToStationary: 0.5
 --- !u!1 &5248060467903245459
 GameObject:
   m_ObjectHideFlags: 0
@@ -832,53 +783,6 @@ MonoBehaviour:
   m_MinIntensity: 6
   m_MaxIntensity: 12
   m_Smoothing: 36
---- !u!1 &8756950943414709585
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 4150610526904978245}
-  - component: {fileID: 5143127320103049556}
-  m_Layer: 0
-  m_Name: Nav Mesh ObstacleR
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &4150610526904978245
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8756950943414709585}
-  m_LocalRotation: {x: 0.000000021855696, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 7980115660496427320}
-  m_RootOrder: 8
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!208 &5143127320103049556
-NavMeshObstacle:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8756950943414709585}
-  m_Enabled: 1
-  serializedVersion: 3
-  m_Shape: 1
-  m_Extents: {x: 1.065, y: 3.4370434, z: 1.29}
-  m_MoveThreshold: 0.1
-  m_Carve: 1
-  m_CarveOnlyStationary: 1
-  m_Center: {x: 7.5, y: 2.187421, z: -0.006958308}
-  m_TimeToStationary: 0.5
 --- !u!1001 &1191869150735884341
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -934,7 +838,8 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 3372118818997572424, guid: 556028c73d0f66d4da492e6ab380cd1c, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 556028c73d0f66d4da492e6ab380cd1c, type: 3}
 --- !u!4 &9094261077284239629 stripped
 Transform:
@@ -996,7 +901,8 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 3372118818997572424, guid: 556028c73d0f66d4da492e6ab380cd1c, type: 3}
   m_SourcePrefab: {fileID: 100100000, guid: 556028c73d0f66d4da492e6ab380cd1c, type: 3}
 --- !u!4 &8400297789364225683 stripped
 Transform:
