+++ Assets/BossRoom/Prefabs/Dungeon/Dungeon Pieces/env_torch2.prefab
@@ -122,8 +122,7 @@ GameObject:
   - component: {fileID: 445632146649864073}
   - component: {fileID: 7463718357983197496}
   - component: {fileID: 8610266421428556183}
-  - component: {fileID: -5560002897593768966}
-  m_Layer: 0
+  m_Layer: 8
   m_Name: env_torch2
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -209,22 +208,6 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1.9255667, y: 2.9335282, z: 1.9451132}
   m_Center: {x: -0.000000029802322, y: 1.4667641, z: -0.00000020861626}
---- !u!208 &-5560002897593768966
-NavMeshObstacle:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4371491365999748502}
-  m_Enabled: 1
-  serializedVersion: 3
-  m_Shape: 1
-  m_Extents: {x: 0.96300006, y: 1.4670001, z: 0.97300005}
-  m_MoveThreshold: 0.1
-  m_Carve: 1
-  m_CarveOnlyStationary: 1
-  m_Center: {x: 0, y: 1.4670001, z: 0}
-  m_TimeToStationary: 0.5
 --- !u!1 &7628197252956779504
 GameObject:
   m_ObjectHideFlags: 0
