+++ Assets/BossRoom/Prefabs/Dungeon/Dungeon Pieces/env_wall1_base_diagonal_1.prefab
@@ -92,7 +92,6 @@ GameObject:
   m_Component:
   - component: {fileID: 5915529853195630635}
   - component: {fileID: 7898993292670171814}
-  - component: {fileID: 4066017979146867213}
   m_Layer: 8
   m_Name: Diagonal_Collider
   m_TagString: Untagged
@@ -127,19 +126,3 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 4, z: 6}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!208 &4066017979146867213
-NavMeshObstacle:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8750842352150602452}
-  m_Enabled: 1
-  serializedVersion: 3
-  m_Shape: 1
-  m_Extents: {x: 0.5, y: 2, z: 3}
-  m_MoveThreshold: 0.1
-  m_Carve: 1
-  m_CarveOnlyStationary: 1
-  m_Center: {x: 0, y: 0, z: 0}
-  m_TimeToStationary: 0.5
