+++ Assets/BossRoom/Prefabs/Dungeon/Dungeon Pieces/env_wall1_base_straight_3.prefab
@@ -12,7 +12,6 @@ GameObject:
   - component: {fileID: 8239492390622286820}
   - component: {fileID: 1148747288013210730}
   - component: {fileID: 8619486518123324128}
-  - component: {fileID: 3653971443187349092}
   m_Layer: 8
   m_Name: env_wall1_base_straight_3
   m_TagString: Untagged
@@ -96,19 +95,3 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 4, z: 4.001817}
   m_Center: {x: 0, y: 2, z: 1.9993523}
---- !u!208 &3653971443187349092
-NavMeshObstacle:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7708204222138769870}
-  m_Enabled: 1
-  serializedVersion: 3
-  m_Shape: 1
-  m_Extents: {x: 0.5, y: 2, z: 2}
-  m_MoveThreshold: 0.1
-  m_Carve: 1
-  m_CarveOnlyStationary: 1
-  m_Center: {x: 0, y: 2, z: 2}
-  m_TimeToStationary: 0.5
