+++ Assets/BossRoom/Prefabs/Dungeon/Dungeon Pieces/env_wall1_top_straight_1.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 2698721903467665021}
   - component: {fileID: 3792955485091636975}
   - component: {fileID: 2036526519701997008}
-  - component: {fileID: 2616948069966954651}
   m_Layer: 8
   m_Name: env_wall1_top_straight_1
   m_TagString: Untagged
@@ -82,16 +81,3 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!65 &2616948069966954651
-BoxCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 759726440798291530}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 2
-  m_Size: {x: 1.0497041, y: 4.0004387, z: 4.001817}
-  m_Center: {x: 0, y: 1.9997805, z: 1.9993523}
