+++ Assets/BossRoom/Prefabs/Dungeon/Dungeon Pieces/pillar/env_pillar2.prefab
@@ -11,8 +11,7 @@ GameObject:
   - component: {fileID: 7980115660496427320}
   - component: {fileID: 8801324764397981686}
   - component: {fileID: 8184890942973508939}
-  - component: {fileID: 7161696663281720505}
-  m_Layer: 6
+  m_Layer: 8
   m_Name: env_pillar2
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -82,19 +81,3 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!208 &7161696663281720505
-NavMeshObstacle:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1899106884463111546}
-  m_Enabled: 1
-  serializedVersion: 3
-  m_Shape: 1
-  m_Extents: {x: 1.8830001, y: 4.144, z: 1.7550001}
-  m_MoveThreshold: 0.1
-  m_Carve: 1
-  m_CarveOnlyStationary: 1
-  m_Center: {x: 2.002, y: 4.1470003, z: -1.1520001}
-  m_TimeToStationary: 0.5
