+++ Assets/BossRoom/Scenes/BossRoom.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4b9be95c5aa3d6060321e054d92cd00f01cc1604ff019fd095b9cb3c6c676def
+oid sha256:1cc62f8667de4ccf8bf6b4b180408ae4f3f206adf92144646e5c0e7d403c7d7b
 size 45793
