+++ Assets/BossRoom/Scenes/BossRoom/NavMesh-DungeonNavMesh.asset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3c71cf5c11a0c1d0984a781785c7fe2f3f8a50775803061ba38f16f0204f1328
-size 57252
+oid sha256:ca584c5914f8dd9d0511ec928ced90ed66c40422365b061f579cdb7f1cfa68d8
+size 57540
