+++ Assets/BossRoom/Scripts/Client/Effects/ScrollingMaterialUVs.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b1352ed2586fb334b801adf6c79951b7
+guid: 3842231b2b526874fa309f7309518e5e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
